# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Timesheets",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/tsapi/public/"
        }
    ],
    "paths": {
...
        "/api/users/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "operationId": "079e21e1527db84e7c0d64a497665796",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "email": "teste@teste.com",
                                    "password": "examplepassword123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/users/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User registrarion",
                "operationId": "047a1334207ff7dd0cfcf0351a72ac8d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "name": "User name",
                                    "email": "teste@teste.com",
                                    "password": "examplepassword123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/users": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "operationId": "4a98e084dce563eefbb013860a62262c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "name": "Michael Kane"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}